requirements.txt
Django==3.0.7
asgiref==3.2.7
pytz==2019.3
sqlparse==0.3.1
mysqlclient
django-crispy-forms
Pillow
djangorestframework
